refactor:  provider에서 apiBaseUrl 필드를 제거하고 기타 설정만 유지하도록 변경
